@@ -40,12 +40,12 @@
     "@kkt/less-modules": "^7.4.9",
     "@kkt/raw-modules": "^7.4.9",
     "@kkt/scope-plugin-options": "^7.4.9",
-    "@types/react": "~17.0.39",
-    "@types/react-dom": "~17.0.13",
+    "@types/react": "^18.0.33",
+    "@types/react-dom": "^18.0.11",
     "@uiw/react-github-corners": "~1.5.3",
     "@uiw/react-markdown-preview": "~3.5.1",
-    "react": "~17.0.2",
-    "react-dom": "~17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "kkt": "^7.4.9",
     "tsbb": "^4.1.4"
   },
